generate_psa_tests: comment out always-skipped test cases
When we generate a test case for a mechanism that is not implemented,
comment out the test case rather than giving it a never-fulfilled
dependency. That way we don't create test cases that cannot be executed.

This changes the generated output in the following ways:

* No longer emit test cases with a dependency on
  `DEPENDENCY_NOT_IMPLEMENTED_YET`. All removed lines that start with
  `depends_on:` contain `DEPENDENCY_NOT_IMPLEMENTED_YET.
* Emit commented-out test cases instead: all the new lines are comment lines.

There is no change in which test cases actually get executed. This removes
many test cases from the list of available test cases, which causes some of
the exceptions in `analyze_outcomes.py` to no longer be useful.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>