PSA test case generation: operation fail: skip never-implemented mechanisms
In `OpFail` test cases, remove the temporary hack whereby test cases were
not skipped when they should be due to a mechanism being never implemented.

This changes many test cases in
`test_suite_psa_crypto_op_fail.generated.data` to be commented out with a
"skipped because" reason instead of having a dependency on an algorithm or
an ECC/DH group that is not implemented.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>