Fix fatal exception on expected failures
Use sets to track membership rather than lists. It's more natural.

In test_compliance, don't overwrite the mutable list passed by the caller.
It's rude.

This commit fixes a bug whereby an expected failure would lead to code that
raised a `ValueError`, with an exception handler that expected this
condition to raise `KeyError`.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>