Read current options from a shadow file rather than the config file
The list of current options needs to be up-to-date, so we can't just use
historical data.

Users may edit the default config file (`include/tf-psa-crypto/crypto_config.h`
or `include/mbedtls/mbedtls_config.h`). One of the reasons we use the list of
config options is to prevent users from defining internal macros in their
config file (in code generated by `generated_config_checks.py`). An internal
macro is one that isn't listed in the official config file. So we need to
know what macros are listed in the official config file, regardless of edits
to the current config file.

Hence we read a "shadow file" which contains the official list of options,
rather than the "live" config file. This file will need to be kept
up-to-date during development, but is not user-editable.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>